# docs/swagger.yaml
basePath: /api
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  routes.DeleteAccountRequest:
    properties:
      confirm:
        type: string
      password:
        type: string
    required:
    - confirm
    - password
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Secure digital banking platform with JWT authentication, TOTP 2FA,
    multi-currency wallets, atomic transfers, and admin monitoring
  title: SecureWallet API
  version: 1.0.0
paths:
...
  /admin/users:
    get:
      consumes:
      - application/json
      description: Get all users for admin management
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Admin list users
      tags:
      - admin
...
  /users:
    get:
      consumes:
      - application/json
      description: Get all users with pagination (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/gin.H'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user account (admin only)
      parameters:
      - description: User creation data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user account (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get a specific user by ID (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update a user's attributes (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User update data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /users/account:
    delete:
      consumes:
      - application/json
      description: Delete the authenticated user's own account after password verification
      parameters:
      - description: Account deletion confirmation
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/routes.DeleteAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Delete own account
      tags:
      - users
  /users/search:
    get:
      consumes:
      - application/json
      description: Search for users by username or email (authenticated users only)
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/gin.H'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Search users
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: Enter the bearer token string with the "Bearer " prefix, e.g. "Bearer
      eyJhbGciOiJIUzI1NiIs..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"